+++ R/data-source-mzmine.R
@@ -26,7 +26,7 @@
 #' ## Create a minimal MZmine 2 feature list CSV
 #' feature_list_path <- file.path(tmp_dir, "mzmine_features.csv")
 #'
-#' mzmine_features <- data.frame(
+#' mzmine_features <- tibble::tibble(
 #'   "row m/z" = c(100.1, 200.2),
 #'   "sample1.mzML Feature status" = c("DETECTED", "DETECTED"),
 #'   "sample1.mzML Feature m/z" = c(100.1, 200.2),
@@ -142,6 +142,7 @@ MZmineFeatureListsSource <- function(
     })
 
     peaks <- do.call(rbind, peaks_list) |>
+        as_tibble() |>
         convert_rt_to_seconds()
 
     new(