+++ R/zzz.R
@@ -1,6 +1,7 @@
 #' @importFrom methods show new validObject slot slotNames is as
 #' @importFrom stats as.formula setNames
 #' @import dplyr
+#' @import tibble
 #' @import ggplot2
 NULL